move branch tripping to util
