+++ src/main/java/com/powsybl/openloadflow/util/BranchTripping.java
@@ -4,7 +4,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package com.powsybl.openloadflow.sa;
+package com.powsybl.openloadflow.util;
 
 import com.powsybl.commons.PowsyblException;
 import com.powsybl.computation.ComputationManager;
