+++ src/test/java/com/powsybl/openloadflow/sa/OpenSecurityAnalysisTest.java
@@ -12,9 +12,6 @@ import com.powsybl.contingency.BranchContingency;
 import com.powsybl.contingency.ContingenciesProvider;
 import com.powsybl.contingency.Contingency;
 import com.powsybl.iidm.network.Network;
-import com.powsybl.iidm.network.Substation;
-import com.powsybl.iidm.network.TopologyKind;
-import com.powsybl.iidm.network.VoltageLevel;
 import com.powsybl.iidm.network.test.EurostagTutorialExample1Factory;
 import com.powsybl.iidm.network.test.FourSubstationsNodeBreakerFactory;
 import com.powsybl.loadflow.LoadFlowParameters;
@@ -25,6 +22,7 @@ import com.powsybl.openloadflow.graph.NaiveGraphDecrementalConnectivity;
 import com.powsybl.openloadflow.network.LfBus;
 import com.powsybl.openloadflow.network.MostMeshedSlackBusSelector;
 import com.powsybl.openloadflow.network.NameSlackBusSelector;
+import com.powsybl.openloadflow.network.NodeBreakerNetworkFactory;
 import com.powsybl.security.LimitViolationFilter;
 import com.powsybl.security.Security;
 import com.powsybl.security.SecurityAnalysisParameters;
@@ -47,160 +45,9 @@ class OpenSecurityAnalysisTest {
 
     private Network network;
 
-    /**
-     *                   G
-     *              C    |
-     * BBS1 -------[+]------- BBS2     VL1
-     *         |        [+] B1
-     *         |         |
-     *     L1  |         | L2
-     *         |         |
-     *     B3 [+]       [+] B4
-     * BBS3 -----------------          VL2
-     *             |
-     *             LD
-     *
-     * 6 buses
-     * 6 branches
-     *
-     *            G
-     *            |
-     *      o--C--o
-     *      |     |
-     *      |     B2
-     *      |     |
-     *      |     o
-     *      |     |
-     *      L1    L2
-     *      |     |
-     *      o     o
-     *      |     |
-     *      B3    B4
-     *      |     |
-     *      ---o---
-     *         |
-     *         LD
-     */
-    static Network createNetwork() {
-        Network network = Network.create("test", "test");
-        Substation s = network.newSubstation()
-                .setId("S")
-                .add();
-        VoltageLevel vl1 = s.newVoltageLevel()
-                .setId("VL1")
-                .setNominalV(400)
-                .setLowVoltageLimit(370.)
-                .setHighVoltageLimit(420.)
-                .setTopologyKind(TopologyKind.NODE_BREAKER)
-                .add();
-        vl1.getNodeBreakerView().newBusbarSection()
-                .setId("BBS1")
-                .setNode(0)
-                .add();
-        vl1.getNodeBreakerView().newBusbarSection()
-                .setId("BBS2")
-                .setNode(1)
-                .add();
-        vl1.getNodeBreakerView().newBreaker()
-                .setId("C")
-                .setNode1(0)
-                .setNode2(1)
-                .add();
-        vl1.getNodeBreakerView().newBreaker()
-                .setId("B1")
-                .setNode1(1)
-                .setNode2(3)
-                .add();
-        vl1.getNodeBreakerView().newInternalConnection()
-                .setNode1(1)
-                .setNode2(4)
-                .add();
-        vl1.getNodeBreakerView().newInternalConnection()
-                .setNode1(0)
-                .setNode2(5)
-                .add();
-        vl1.newGenerator()
-                .setId("G")
-                .setNode(4)
-                .setMinP(-9999.99)
-                .setMaxP(9999.99)
-                .setVoltageRegulatorOn(true)
-                .setTargetV(398)
-                .setTargetP(603.77)
-                .setTargetQ(301.0)
-                .add();
-
-        VoltageLevel vl2 = s.newVoltageLevel()
-                .setId("VL2")
-                .setNominalV(400)
-                .setTopologyKind(TopologyKind.NODE_BREAKER)
-                .setLowVoltageLimit(370.)
-                .setHighVoltageLimit(420.)
-                .add();
-        vl2.getNodeBreakerView().newBusbarSection()
-                .setId("BBS3")
-                .setNode(0)
-                .add();
-        vl2.getNodeBreakerView().newBreaker()
-                .setId("B3")
-                .setNode1(0)
-                .setNode2(1)
-                .add();
-        vl2.getNodeBreakerView().newBreaker()
-                .setId("B4")
-                .setNode1(0)
-                .setNode2(2)
-                .add();
-        vl2.getNodeBreakerView().newInternalConnection()
-                .setNode1(0)
-                .setNode2(3)
-                .add();
-        vl2.newLoad()
-                .setId("LD")
-                .setNode(3)
-                .setP0(600.0)
-                .setQ0(200.0)
-                .add();
-
-        network.newLine()
-                .setId("L1")
-                .setVoltageLevel1("VL1")
-                .setNode1(5)
-                .setVoltageLevel2("VL2")
-                .setNode2(1)
-                .setR(3.0)
-                .setX(33.0)
-                .setG1(0.0)
-                .setB1(386E-6 / 2)
-                .setG2(0.0)
-                .setB2(386E-6 / 2)
-                .add();
-
-        network.newLine()
-                .setId("L2")
-                .setVoltageLevel1("VL1")
-                .setNode1(3)
-                .setVoltageLevel2("VL2")
-                .setNode2(2)
-                .setR(3.0)
-                .setX(33.0)
-                .setG1(0.0)
-                .setB1(386E-6 / 2)
-                .setG2(0.0)
-                .setB2(386E-6 / 2)
-                .add();
-
-        network.getLine("L1").newCurrentLimits1().setPermanentLimit(940.0).add();
-        network.getLine("L1").newCurrentLimits2().setPermanentLimit(940.0).add();
-        network.getLine("L2").newCurrentLimits1().setPermanentLimit(940.0).add();
-        network.getLine("L2").newCurrentLimits2().setPermanentLimit(940.0).add();
-
-        return network;
-    }
-
     @BeforeEach
     void setUp() {
-        network = createNetwork();
+        network = NodeBreakerNetworkFactory.create();
     }
 
     @Test
